\name{degrees}
\alias{degrees}

\title{ Radians to degrees }

\description{ Accessory function to transform radians into degrees. }

\usage{degrees(radian)}

\arguments{
  \item{radian}{ Angle in radians and decimal fraction.}
}

\value{Angle in degrees.}

\seealso{\code{\link{radians}}}

\examples{
degrees(seq(0,2*pi,pi/2))
}