\name{hourangle}
\alias{hourangle}

\title{Hour angle}
\description{
Hour angle, internal function for solar position.
}
\usage{
hourangle(jd, longitude, timezone)
}

\arguments{
  \item{jd}{Julian Day.}

  \item{longitude}{
Longitude.
}  
\item{timezone}{
Timezone in hours, west of Greenwich is negative.
}
}

\value{
Hour angle
}

