\name{radians}
\alias{radians}

\title{ Degrees to radians }

\description{ Accessory function to transform degrees into radians. }

\usage{radians(degree)}

\arguments{
  \item{degree}{ Angle in degrees and decimal fraction.}
}

\value{Angle in radians.}

\author{Javier G. Corripio}

\seealso{\code{\link{degrees}}}

\examples{
radians(seq(0,360,90))
}
