\name{hourangle}
\alias{hourangle}

\title{Hour Angle}
\description{
Hour angle, internal function for solar position.
}
\usage{
hourangle(jd, longitude, timezone)
}

\arguments{
  \item{jd}{Julian Day}

  \item{longitude}{
Longitude
}  
\item{timezone}{
timezone in hours, west of Greenwich is negative
}
}

\value{
Hour angle
}

\author{
Javier G. Corripio
}

